Write to serial device when RFID changes

+++ communication/rfid.h
@@ -5,12 +5,12 @@
 #ifndef ARCTOS_ROBOT_RFID_H
 #define ARCTOS_ROBOT_RFID_H
 
-
+#include <unistd.h>
 #include <stddef.h>
 #include "../general.h"
 
 int rfid_init(serial_port_options_t* options);
 
-int rfid_read(int fd, char* buffer, size_t buffer_size);
+ssize_t rfid_read(int fd, char* buffer, size_t buffer_size);
 
 #endif //ARCTOS_ROBOT_RFID_H
